fix: extract SidebarItem and SidebarItmes component from sidebar component
