+++ src/layouts/SideBar.tsx
@@ -1,31 +1,7 @@
 import React, { useContext } from "react";
-import { useRouter } from "next/router";
-import Link from "next/link";
 import { MusicContext } from "@/contexts/MusicContext";
-import { SiYoutubemusic } from "react-icons/si";
-import { IoEllipsisHorizontalSharp } from "react-icons/io5";
-import { FiPlus } from "react-icons/fi";
-import { TbClockPlay } from "react-icons/tb";
-import { MdHomeFilled, MdQueueMusic } from "react-icons/md";
-import { FaUsers } from "react-icons/fa";
-import {
-  RiRocket2Fill,
-  RiAlbumFill,
-  RiFolderMusicFill,
-  RiHeartLine,
-  RiDeleteBinLine,
-} from "react-icons/ri";
-
-interface SidebarCategory {
-  label: string;
-  subCategories: SidebarMenuItem[];
-}
-
-interface SidebarMenuItem {
-  id?: string;
-  label: string;
-  path: string;
-}
+import { iconList } from "@/components/sidebar/SidebarItem";
+import RenderSidebarItems from "@/components/sidebar/SidebarItems";
 
 export const sidebarContent = [
   {
@@ -88,86 +64,6 @@ export const sidebarContent = [
   },
 ];
 
-export const iconList: Record<string, React.ReactNode> = {
-  musicPlayer: <SiYoutubemusic className="h-10 w-10" />,
-  more: <IoEllipsisHorizontalSharp className="ml-[63px] hidden h-8 w-8 lg:block" />,
-  Home: <MdHomeFilled className="h-6 w-6" />,
-  Discover: <RiRocket2Fill className="h-6 w-6" />,
-  Albums: <RiAlbumFill className="h-6 w-6" />,
-  Artists: <FaUsers className="h-6 w-6" />,
-  Videos: <RiFolderMusicFill className="h-6 w-6" />,
-  "Recently Played": <TbClockPlay className="h-6 w-6" />,
-  "Favourite Songs": <RiHeartLine className="h-6 w-6" />,
-  playlist: <MdQueueMusic className="h-6 w-6" />,
-  Plus: <FiPlus className="h-4 w-4" />,
-  Delete: <RiDeleteBinLine className="h-4 w-4" />,
-};
-
-interface RenderSidebarItemProps {
-  isPlaylistsCategory: boolean;
-  menuItem: SidebarMenuItem;
-}
-
-const RenderSidebarItem: React.FC<RenderSidebarItemProps> = ({ isPlaylistsCategory, menuItem }) => {
-  const router = useRouter();
-  return (
-    <Link
-      href={isPlaylistsCategory ? `${menuItem.path}/${menuItem.id}` : menuItem.path}
-      className="group flex items-center"
-    >
-      <div
-        className={`h-8 w-2 ${
-          router.asPath.split("?")[0] ===
-          (isPlaylistsCategory ? `${menuItem.path}/${menuItem.id}` : menuItem.path)
-            ? "bg-primary"
-            : ""
-        }`}
-      ></div>
-      <div
-        className={`flex h-8 items-center pl-4 pr-5 group-hover:text-light group-focus:text-light lg:ml-0 lg:pl-6 ${
-          router.asPath.split("?")[0] ===
-          (isPlaylistsCategory ? `${menuItem.path}/${menuItem.id}` : menuItem.path)
-            ? "bg-gradient-to-r from-primary to-transparent"
-            : ""
-        }`}
-      >
-        {isPlaylistsCategory ? iconList["playlist"] : iconList[menuItem.label]}
-      </div>
-      <li className="flex items-center lg:gap-[51px]">
-        <p className="hidden group-hover:text-light group-focus:text-light lg:block">
-          {menuItem.label}
-        </p>
-        <div className="hidden h-6 w-6 items-center justify-center text-light-100 hover:text-light lg:flex">
-          {isPlaylistsCategory && iconList["Delete"]}
-        </div>
-      </li>
-    </Link>
-  );
-};
-
-const RenderSidebarItems = (category: SidebarCategory) => {
-  const isPlaylistsCategory = category.label === "PLAYLISTS";
-  return (
-    <div key={category.label}>
-      <div className="invisible mb-8 mt-[54px] flex items-center text-light-200 lg:visible lg:ml-[36px] lg:gap-[146px]">
-        <p className="text-primary-100 lg:inline-block lg:whitespace-nowrap">{category.label}</p>
-        <div className="hidden h-6 w-6 items-center justify-center text-light-300 hover:text-light lg:flex">
-          {isPlaylistsCategory && iconList["Plus"]}
-        </div>
-      </div>
-      <ul className="flex flex-col gap-8 font-roboto text-lg text-light-200">
-        {category.subCategories.map((menuItem, subIndex) => (
-          <RenderSidebarItem
-            isPlaylistsCategory={isPlaylistsCategory}
-            menuItem={menuItem}
-            key={subIndex}
-          />
-        ))}
-      </ul>
-    </div>
-  );
-};
-
 const SideBar = () => {
   const { selectedAlbum } = useContext(MusicContext);
   const musicContext = selectedAlbum?.musicContext;
@@ -186,7 +82,7 @@ const SideBar = () => {
           {iconList["more"]}
         </div>
         <div className="overflow-y-auto overflow-x-hidden pb-12">
-          {sidebarContent.map((category, index) => RenderSidebarItems(category))}
+          {sidebarContent.map((category) => RenderSidebarItems(category))}
         </div>
       </div>
     </aside>
